requirements.txt
Django==3.0
requests==2.9.1
python-magic==0.4.15
Pillow==6.2.1
psycopg2-binary==2.8.4
dj-database-url==0.5.0
djangorestframework==3.11.0
